Systems/amazon-growth-os/requirements.txt
# Amazon Growth OS Dependencies
# Version: 1.0

# Core AI/ML Framework
crewai>=1.7.0
crewai-tools>=1.7.2
litellm>=1.75.0
anthropic>=0.75.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
duckdb>=1.4.0

# Vector Database & Embeddings
qdrant-client>=1.16.0
sentence-transformers>=5.0.0

# Web Framework (Dashboard)
streamlit>=1.50.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0

# Optional: For API integrations
# boto3>=1.34.0  # AWS/Amazon APIs
# google-cloud-storage>=2.0.0  # GCS backup
